Change file substitution for temporary file
The `make test` target, although working on a
local machine, did not work on the Travis-CI
build: it fails at the first test of the `testTty`
test from `tests/rperr.bash`.

The same behavior was reproduced on the local
machine when the `source` line was remove, so a
possible explanation is that Bash's file
substitution (`<(cmd)`) is not acting correctly on
the Travis-CI container.

Change the `source` command argument and some
previous line, so that a temporary file is sourced
instead.