Remove `rperr.bash`
The script when run the Travis-CI container fails
the following condition `if tty`, which means that
the `rperr` statement that is run is the one for
when the script is running on a system with an
x-server and not from the terminal.

By removing the aforementioned test script, that
Travis-CI build should succeed.