Change `gpg2` call in `tests/scripts/context.bash`
As it is, the \`gpg2\` call needs access to a tty,
which is fine on most local machines, but not the
Travis-CI container, making the build fail at such
point.

Change the `gpg2` call to make use of the
`--command-file` instead of the `--command-fd`
option, to see if the build will succeed.