Fix `declare -x .*` being printed to stdout
Whenever a context is used, if successful, the
`source` command will cause all exported variables
to be printed to stdout. This behavior is caused
by the sed script in `get_context()` not deleting
empty lines of the read file.

Fix this by adding empty deletion for the
aforementioned sed script.